Add change state column to diff tree view
