+++ src/NexusMods.App.UI/Controls/Trees/Files/IFileTreeNodeViewModel.cs
@@ -2,6 +2,9 @@ using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.Loadouts.Synchronizers;
 using NexusMods.App.UI.Controls.Trees.Common;
 using NexusMods.App.UI.Helpers.TreeDataGrid;
+using NexusMods.App.UI.Resources;
+using NexusMods.App.UI.Resources;
+using NexusMods.App.UI.Resources;
 
 namespace NexusMods.App.UI.Controls.Trees.Files;
 
@@ -11,27 +14,27 @@ public interface IFileTreeNodeViewModel : IViewModelInterface, IExpandableItem, 
     ///     The icon that's used to display this specific node.
     /// </summary>
     FileTreeNodeIconType Icon { get; set; }
-    
+
     /// <summary>
     ///     True if this node represents a file.
     /// </summary>
     bool IsFile { get; }
-    
+
     /// <summary>
     ///     Name of the file or folder segment.
     /// </summary>
     string Name { get; }
-    
+
     /// <summary>
     ///     The size of the file, in bytes.
     /// </summary>
     ulong FileSize { get; }
-    
+
     /// <summary>
     ///     Total number of files descending from this node.
     /// </summary>
     uint FileCount { get; }
-    
+
     /// <summary>
     ///     The change status of the file (for diff views).
     /// </summary>
@@ -44,6 +47,19 @@ public interface IFileTreeNodeViewModel : IViewModelInterface, IExpandableItem, 
     string ToFormattedFileCount()
     {
         return FileCount > 0 ? FileCount.ToString() : string.Empty;
+    }
+
+    string ToFormattedChangeState()
+    {
+        return ChangeType switch
+        {
+            FileChangeType.Added => Language.IFileTreeNodeViewModel_ToFormattedChangeState_Added,
+            FileChangeType.Modified => IsFile
+                ? Language.IFileTreeNodeViewModel_ToFormattedChangeState_Modified
+                : Language.IFileTreeNodeViewModel_ToFormattedChangeState_Contents_modified,
+            FileChangeType.Removed => Language.IFileTreeNodeViewModel_ToFormattedChangeState_Removed,
+            FileChangeType.None => string.Empty,
+        };
     }
 
     /// <summary>
