+++ src/NexusMods.App.UI/Resources/Language.pl.resx
@@ -353,7 +353,7 @@
         <value>Pliki: {0} ({1})</value>
         <comment>{0} Represents a number, {1} represents a formatted size (e.g. 12GB)</comment>
     </data>
-    <data name="ModFileTreeViewModel_CreateTreeSource_File_Count" xml:space="preserve">
+    <data name="Helpers_GenerateHeader_File_Count" xml:space="preserve">
         <value>ILOŚĆ PLIKÓW</value>
     </data>
     <data name="IngestButtonTextBlock_Text" xml:space="preserve">
